All reuseable methods moved to utility.h
