+++ firstFit.h
@@ -3,4 +3,5 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include "utility.h"
 #include "memoryStructure.h"
